# openapi_spec.yaml
openapi: 3.0.3
info:
  title: Docaiche API
  description: |
    AI-powered documentation search and content management system with advanced analytics capabilities.
    
    ## Overview
    Docaiche provides intelligent documentation search, content enrichment, and comprehensive analytics
    for tracking user behavior and system performance. The API supports real-time search with caching,
    content ingestion from multiple sources, and AI-powered content enhancement.
    
    ## Key Features
    - **Intelligent Search**: AI-powered search with technology filtering and relevance scoring
    - **Content Management**: Ingest and manage documentation from various sources
    - **Analytics Dashboard**: Comprehensive metrics for search patterns, user behavior, and system performance
    - **AI Enrichment**: Automatic content enhancement using LLM providers
    - **Real-time Monitoring**: Live system health and performance metrics
    
    ## Analytics Capabilities
    The analytics endpoints provide rich data for building beautiful dashboards:
    - Search query trends and patterns over time
    - User engagement metrics (clicks, dwell time, scroll depth)
    - Content quality scores and performance
    - System performance metrics (response times, cache hit rates)
    - Technology adoption trends
    - Real-time activity feeds
    
    ## Monitoring Integration
    - **Loki**: Centralized log aggregation for all services
...
    
  version: 2.0.0
  contact:
...
  license:
    name: MIT
...
servers:
...
paths:
...
  /stats:
    get:
      tags: [health, analytics]
      summary: System statistics
      description: |
        Provides comprehensive usage and performance statistics.
        
        **Use Cases for Dashboards:**
        - Real-time performance monitoring widgets
        - Cache efficiency gauges
        - Search volume time series charts
        - System resource utilization graphs
        
        **Metrics Included:**
        - Search statistics: total searches, response times, cache performance
        - Content statistics: document counts, quality scores, workspace distribution
        - System statistics: CPU, memory, disk usage, uptime
        - Cache statistics: hit/miss rates, memory usage, evictions
        
        **Note:** Currently returns real statistics from the database when available,
        with graceful fallback to empty/zero values if tables are not yet populated.
      operationId: getStats
      responses:
        '200':
          description: System statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsResponse'
        '500':
          description: Statistics unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /analytics:
    get:
      tags: [analytics]
      summary: Detailed analytics data
      description: |
        Returns comprehensive analytics for the specified time range.
        
        **Dashboard Visualization Ideas:**
        - Search query word clouds from top queries
        - Technology adoption pie charts
        - User engagement funnel charts
        - Time series graphs for all metrics
        - Geographic distribution maps (if user location data available)
        
        **Data Granularity:**
        - Supports 24h, 7d, and 30d time ranges
        - Returns aggregated metrics suitable for visualization
        - Includes top-N lists for queries and technologies
        
        **Current Status:** Returns demonstration data showing the structure
        and types of analytics available. Will return real data as the system
        accumulates usage metrics.
      operationId: getAnalytics
      parameters:
        - name: timeRange
          in: query
          description: Time range for analytics
          required: false
          schema:
            type: string
            enum: [24h, 7d, 30d]
            default: 24h
      responses:
        '200':
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsResponse'
...
components:
  schemas:
...
    StatsResponse:
      type: object
      required: [search_stats, cache_stats, content_stats, system_stats, timestamp]
      properties:
        search_stats:
          type: object
          properties:
            total_searches:
              type: integer
              description: Total number of searches performed
            avg_response_time_ms:
              type: number
              description: Average search response time in milliseconds
            cache_hit_rate:
              type: number
              description: Percentage of searches served from cache
            successful_searches:
              type: integer
              description: Number of successful searches
            failed_searches:
              type: integer
              description: Number of failed searches
        cache_stats:
          type: object
          properties:
            hit_rate:
              type: number
              description: Cache hit rate (0-1)
            miss_rate:
              type: number
              description: Cache miss rate (0-1)
            total_keys:
              type: integer
              description: Total keys in cache
            memory_usage_mb:
              type: number
              description: Cache memory usage in MB
            evictions:
              type: integer
              description: Number of cache evictions
        content_stats:
          type: object
          properties:
            total_documents:
              type: integer
              description: Total documents in system
            total_chunks:
              type: integer
              description: Total document chunks
            avg_quality_score:
              type: number
              description: Average document quality score (0-1)
            workspaces:
              type: integer
              description: Number of workspaces
            last_enrichment:
              type: string
              format: date-time
              description: Last enrichment timestamp
        system_stats:
          type: object
          properties:
            uptime_seconds:
              type: integer
              description: System uptime in seconds
            cpu_usage_percent:
              type: number
              description: CPU usage percentage
            memory_usage_mb:
              type: number
              description: Memory usage in MB
            disk_usage_mb:
              type: number
              description: Disk usage in MB
        timestamp:
          type: string
          format: date-time
    AnalyticsResponse:
      type: object
      required: [timeRange, searchMetrics, contentMetrics, userMetrics, timestamp]
      properties:
        timeRange:
          type: string
          enum: [24h, 7d, 30d]
        searchMetrics:
          type: object
          properties:
            totalSearches:
              type: integer
              description: Total searches in period
            avgResponseTime:
              type: number
              description: Average response time
            successRate:
              type: number
              description: Search success rate (0-1)
            topQueries:
              type: array
              items:
                type: object
                properties:
                  query:
                    type: string
                  count:
                    type: integer
        contentMetrics:
          type: object
          properties:
            totalDocuments:
              type: integer
            totalChunks:
              type: integer
            avgQualityScore:
              type: number
            documentsByTechnology:
              type: array
              items:
                type: object
                properties:
                  technology:
                    type: string
                  count:
                    type: integer
        userMetrics:
          type: object
          properties:
            activeUsers:
              type: integer
            totalSessions:
              type: integer
            avgSessionDuration:
              type: number
              description: Average session duration in milliseconds
        timestamp:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: Error message
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (future implementation)
security: []  # No security currently implemented